Remove tree from public API of v2.0.0 (duplication with SGMap)
